+++ Meltdown/Items/White/PlutoniumRounds.cs
@@ -1,8 +1,11 @@
-﻿using Meltdown.Orbs;
+﻿using HG;
+using Meltdown.Orbs;
 using Meltdown.Utils;
 using R2API;
 using RoR2;
+using RoR2.Items;
 using UnityEngine;
+using UnityEngine.Networking;
 
 namespace Meltdown.Items.White
 {
@@ -25,46 +28,100 @@ namespace Meltdown.Items.White
             return ItemDisplayRuleUtils.getPlutoniumRoundsDisplay(prefab);
         }
 
-        public override void Hooks()
+        public override void Hooks() { }
+
+        public void FirePlutoniumRound(CharacterBody body)
         {
-            On.RoR2.CharacterBody.OnSkillActivated += CharacterBody_OnSkillActivated;
+            var stack = GetCount(body);
+            var radius = 25 + (5 * stack);
+
+            HurtBox[] hurtBoxes = new SphereSearch
+            {
+                origin = body.transform.position,
+                radius = radius,
+                mask = LayerIndex.entityPrecise.mask,
+                queryTriggerInteraction = QueryTriggerInteraction.UseGlobal
+            }.RefreshCandidates().FilterCandidatesByHurtBoxTeam(TeamMask.GetEnemyTeams(body.teamComponent.teamIndex)).OrderCandidatesByDistance().FilterCandidatesByDistinctHurtBoxEntities().GetHurtBoxes();
+
+            for (int i = 0; i < Mathf.Min(stack, hurtBoxes.Length); i++)
+            {
+                IrradiatedOrb orb = new (body.gameObject, body.damage * 2.0f, body.RollCrit(), body.transform.position, body.teamComponent.teamIndex, hurtBoxes[i]);
+                RoR2.Orbs.OrbManager.instance.AddOrb(orb);
+            }
+        }
+    }
+
+    public class PlutoniumRoundsController : BaseItemBodyBehavior
+    {
+        [ItemDefAssociation]
+        private static ItemDef GetItemDef() => Meltdown.items.plutoniumRounds.itemDef;
+
+        private readonly float itemInterval = 0.5f;
+        private float itemSecondaryCooldown = 0.0f;
+        private float itemUtilityCooldown = 0.0f;
+        private float itemSpecialCooldown = 0.0f;
+
+        public void OnEnable()
+        {
+            itemSecondaryCooldown = 0.0f;
+            itemUtilityCooldown = 0.0f;
+            itemSpecialCooldown = 0.0f;
+            body.onSkillActivatedServer += Body_onSkillActivatedServer;
         }
 
-        private void CharacterBody_OnSkillActivated(On.RoR2.CharacterBody.orig_OnSkillActivated orig, CharacterBody self, GenericSkill skill)
+        public void OnDisable()
         {
-            orig(self, skill);
+            body.onSkillActivatedServer -= Body_onSkillActivatedServer;
+        }
 
-            if (!self || skill == null || !self.inventory)
+        public void FixedUpdate()
+        {
+            if (!NetworkServer.active)
             {
                 return;
             }
 
-            var stack = GetCount(self);
-            var skillLocator = self.GetComponent<SkillLocator>();
+            itemSecondaryCooldown += Time.fixedDeltaTime;
+            itemUtilityCooldown += Time.fixedDeltaTime;
+            itemSpecialCooldown += Time.fixedDeltaTime;
+        }
+
+        private void Body_onSkillActivatedServer(GenericSkill skill)
+        {
+            if (skill == null)
+            {
+                return;
+            }
+
+            var stack = Meltdown.items.plutoniumRounds.GetCount(body);
 
             var isRailgunnerScopedPrimary =
-                self.bodyIndex == BodyCatalog.SpecialCases.RailGunner() &&
-                skill == skillLocator.primary &&
-                self.canAddIncrasePrimaryDamage;
+                body.bodyIndex == BodyCatalog.SpecialCases.RailGunner() &&
+                skill == body.skillLocator.primary &&
+                body.canAddIncrasePrimaryDamage;
 
-            var isNonPrimary = skill != skillLocator.primary && skill.cooldownRemaining > 0;
+            var isNonPrimary = skill != body.skillLocator.primary && skill.baseRechargeInterval > 0.0f;
+
+            var secondaryOffCooldown = skill == body.skillLocator.secondary && itemSecondaryCooldown > itemInterval;
+            var utilityOffCooldown = skill == body.skillLocator.utility && itemUtilityCooldown > itemInterval;
+            var specialOffCooldown = skill == body.skillLocator.special && itemSpecialCooldown > itemInterval;
 
             if (stack > 0 && (isNonPrimary || isRailgunnerScopedPrimary))
             {
-                var radius = 25 + (5 * stack);
-
-                HurtBox[] hurtBoxes = new SphereSearch
+                if (secondaryOffCooldown)
                 {
-                    origin = self.transform.position,
-                    radius = radius,
-                    mask = LayerIndex.entityPrecise.mask,
-                    queryTriggerInteraction = QueryTriggerInteraction.UseGlobal
-                }.RefreshCandidates().FilterCandidatesByHurtBoxTeam(TeamMask.GetEnemyTeams(self.teamComponent.teamIndex)).OrderCandidatesByDistance().FilterCandidatesByDistinctHurtBoxEntities().GetHurtBoxes();
-
-                for (int i = 0; i < Mathf.Min(stack, hurtBoxes.Length); i++)
+                    itemSecondaryCooldown = 0.0f;
+                    Meltdown.items.plutoniumRounds.FirePlutoniumRound(body);
+                }
+                else if (utilityOffCooldown)
                 {
-                    IrradiatedOrb orb = new IrradiatedOrb(self.gameObject, self.damage * 2.5f, self.RollCrit(), self.transform.position, self.teamComponent.teamIndex, hurtBoxes[i]);
-                    RoR2.Orbs.OrbManager.instance.AddOrb(orb);
+                    itemUtilityCooldown = 0.0f;
+                    Meltdown.items.plutoniumRounds.FirePlutoniumRound(body);
+                }
+                else if (specialOffCooldown)
+                {
+                    itemSpecialCooldown = 0.0f;
+                    Meltdown.items.plutoniumRounds.FirePlutoniumRound(body);
                 }
             }
         }