+++ Thunderstore/CHANGELOG.md
@@ -1,3 +1,26 @@
+## 0.3.3
+- Reduced the damage bonus of Uranium Fuel Rods to 200%.
+	- A chance on hit and extended dot timer made this item a little too good.
+- Plutonium Rounds and Reactor Vent now have a 0.5s internal cooldown between activations.
+	- Plutonium rounds has a seperate cooldown per skill, so you can still fire it very quickly if you go between skills.
+- Changed the nuclear elite blast attack to a slow on-hit, instead of applying irradiated (all other attacks still apply irradiated).
+- Changed the duration of irradiated applied by nuclear elites to scale down based on proc coefficient.
+	- The upper cap is still 4 seconds, even if the enemy's proc coefficient is greater than 1.
+- Fixed abandonment primary damage scaling.
+- Fixed plutonium rounds, reactor vents and old exhaust pipe not working on non-host clients.
+- Fixed plutonium rounds doing more damage than intended.
+
+## 0.3.2
+- Updated readme
+
+## 0.3.1
+- Reduced irradiated damage per tick from 75% to 50%.
+	- With the introduction of the 5% chance on hit for the green items, this DoT became more common than intended.
+- Increased time between each nuclear elite blast to 6 seconds.
+- Reduced nuclear elite blast attack radius by 20%.
+- Reduced irradiated duration to 4s for nuclear elites only.
+	- For a tier 1 elite, this elite was capable of too much damage out of nowhere. This first pass should help reduce it's lethality, but I have more changes in mind if this isn't enough.
+
 ## 0.3.0
 - Replaced Metal Claws with Thermite-in-a-Can.
 	- This is functionally the same item, but replaces the bleed effect with fire.
@@ -21,9 +44,9 @@
 *Thanks RoR2 Modding discord for the solid feedback on it all so far.*
 
 ## 0.1.1
-- Removed in-dev shenanigans
+- Removed in-dev shenanigans.
 
 ## 0.1.0
 
-- First release
+- First release.
 - Hooray!